+++ src/main/java/com/dextrus/service/FolderService.java
@@ -48,7 +48,7 @@ public class FolderService {
     }
 
 	public void addFolderOrFile(String type, String name, int parentId) {
-		int count = folderRepository.countOfRowsWithSameNameAndParentId(parentId, name+"%");
+		int count = folderRepository.countOfRowsWithSameNameAndParentId(parentId, name);
 		if(count==0) {
 			Folder folder = new Folder();
 			folder.setName(name);
@@ -64,17 +64,21 @@ public class FolderService {
 			folder.setType(type);
 			folderRepository.save(folder);	
 		}
-		System.out.println(count);
 	}
 
 	@Transactional
-	public void deleteFileOrFolder(int id) {
-		folderRepository.deleteAllByParentId(id);
-		folderRepository.deleteById(id);
+	public String deleteFileOrFolder(int id) {
+		if(folderRepository.issParentNull(id)=="NULL")
+			return "PARENT_NULL";
+		else {
+			folderRepository.deleteAllByParentId(id);
+			folderRepository.deleteById(id);
+			return "DELETED";
+		}
 	}
 
 	public void renameFileOrFolder(int id, String newName) {
-
+		
 	}
